Separate create person/student/teacher methods
