+++ create_student.rb
@@ -1,0 +1,15 @@
+require './app'
+def create_student
+  print 'Age: '
+  age = gets.chomp
+
+  print 'Does student have parent permission? [Y/N]: '
+  parent_permission = gets.chomp.to_s.downcase == 'y'
+
+  print 'Name: '
+  name = gets.chomp
+
+  student = Student.new(age, name, parent_permission: parent_permission)
+  @people.push(student)
+  puts 'Student created successfully'
+end
